chore(tracing): ensures trace sampling can only be configured via envars and remote config
