+++ ddtrace/_trace/processor/__init__.py
@@ -9,6 +9,7 @@ from typing import Optional
 from typing import Union
 
 from ddtrace import config
+from ddtrace._trace.sampler import BaseSampler
 from ddtrace._trace.span import Span
 from ddtrace._trace.span import _get_64_highest_order_bits_as_hex
 from ddtrace._trace.span import _is_top_level
@@ -27,7 +28,6 @@ from ddtrace.internal.service import ServiceStatusError
 from ddtrace.internal.telemetry.constants import TELEMETRY_LOG_LEVEL
 from ddtrace.internal.telemetry.constants import TELEMETRY_NAMESPACE_TAG_TRACER
 from ddtrace.internal.writer import TraceWriter
-from ddtrace.sampler import BaseSampler
 
 
 try:
