+++ ddtrace/appsec/_iast/_overhead_control_engine.py
@@ -8,11 +8,11 @@ from typing import Text
 from typing import Tuple
 from typing import Type
 
+from ddtrace._trace.sampler import RateSampler
 from ddtrace._trace.span import Span
 from ddtrace.appsec._iast._utils import _is_iast_debug_enabled
 from ddtrace.internal._unpatched import _threading as threading
 from ddtrace.internal.logger import get_logger
-from ddtrace.sampler import RateSampler
 from ddtrace.settings.asm import config as asm_config
