+++ tests/integration/test_debug.py
@@ -12,11 +12,11 @@ import mock
 import pytest
 
 import ddtrace
+import ddtrace._trace.sampler
 from ddtrace._trace.span import Span
 from ddtrace.internal import debug
 from ddtrace.internal.writer import AgentWriter
 from ddtrace.internal.writer import TraceWriter
-import ddtrace.sampler
 from tests.subprocesstest import SubprocessTestCase
 from tests.subprocesstest import run_in_subprocess
 
@@ -317,7 +317,7 @@ def test_custom_writer():
 
 def test_different_samplers():
     tracer = ddtrace.Tracer()
-    tracer.configure(sampler=ddtrace.sampler.RateSampler())
+    tracer.configure(sampler=ddtrace._trace.sampler.RateSampler())
     info = debug.collect(tracer)
 
     assert info.get("sampler_type") == "RateSampler"
@@ -325,7 +325,7 @@ def test_different_samplers():
 
 def test_startup_logs_sampling_rules():
     tracer = ddtrace.Tracer()
-    sampler = ddtrace.sampler.DatadogSampler(rules=[ddtrace.sampler.SamplingRule(sample_rate=1.0)])
+    sampler = ddtrace._trace.sampler.DatadogSampler(rules=[ddtrace._trace.sampler.SamplingRule(sample_rate=1.0)])
     tracer.configure(sampler=sampler)
     f = debug.collect(tracer)
 
@@ -334,8 +334,8 @@ def test_startup_logs_sampling_rules():
         " tags='NO_RULE', provenance='default')"
     ]
 
-    sampler = ddtrace.sampler.DatadogSampler(
-        rules=[ddtrace.sampler.SamplingRule(sample_rate=1.0, service="xyz", name="abc")]
+    sampler = ddtrace._trace.sampler.DatadogSampler(
+        rules=[ddtrace._trace.sampler.SamplingRule(sample_rate=1.0, service="xyz", name="abc")]
     )
     tracer.configure(sampler=sampler)
     f = debug.collect(tracer)
